requirements.txt
keras==2.1.5
tensorflow==1.6.0
matplotlib
numpy
Pillow
tensorboard
termcolor
google-cloud-storage
setuptools
opencv-python
SciPy